Update translations for new plurals
For languages with only one plural form, I've renamed the placeholders
and removed the "fuzzy" mark from all updated translations in the .po
files. Translators should not need to do anything for these at the
next release.

For languages with multiple plural forms:

- For strings with changed placeholders that were already marked with
ngettext, I've just updated the placeholders in the translations and
removed the fuzzy marks. Translators should not need to do anything
for these.

- For strings newly marked with ngettext, I updated the placeholders
but left the translations marked fuzzy, so Weblate should show them as
needing editing, and translators will have a chance to correct the new
plural versions.